+++ cli/src/main/java/bisq/cli/GrpcStubs.java
@@ -1,0 +1,54 @@
+/*
+ * This file is part of Bisq.
+ *
+ * Bisq is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or (at
+ * your option) any later version.
+ *
+ * Bisq is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package bisq.cli;
+
+import bisq.proto.grpc.GetVersionGrpc;
+import bisq.proto.grpc.OffersGrpc;
+import bisq.proto.grpc.PaymentAccountsGrpc;
+import bisq.proto.grpc.WalletsGrpc;
+
+import io.grpc.CallCredentials;
+import io.grpc.ManagedChannelBuilder;
+
+import static java.util.concurrent.TimeUnit.SECONDS;
+
+public class GrpcStubs {
+
+    public final GetVersionGrpc.GetVersionBlockingStub versionService;
+    public final OffersGrpc.OffersBlockingStub offersService;
+    public final PaymentAccountsGrpc.PaymentAccountsBlockingStub paymentAccountsService;
+    public final WalletsGrpc.WalletsBlockingStub walletsService;
+
+    public GrpcStubs(String apiHost, int apiPort, String apiPassword) {
+        CallCredentials credentials = new PasswordCallCredentials(apiPassword);
+
+        var channel = ManagedChannelBuilder.forAddress(apiHost, apiPort).usePlaintext().build();
+        Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+            try {
+                channel.shutdown().awaitTermination(1, SECONDS);
+            } catch (InterruptedException ex) {
+                throw new IllegalStateException(ex);
+            }
+        }));
+
+        this.versionService = GetVersionGrpc.newBlockingStub(channel).withCallCredentials(credentials);
+        this.offersService = OffersGrpc.newBlockingStub(channel).withCallCredentials(credentials);
+        this.paymentAccountsService = PaymentAccountsGrpc.newBlockingStub(channel).withCallCredentials(credentials);
+        this.walletsService = WalletsGrpc.newBlockingStub(channel).withCallCredentials(credentials);
+    }
+}
